+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Features
 
-* let users report other users ([b12544f](https://adl.github.com/adl-developer/kinkane-backend/commit/b12544f93a793142a3c63cd0bd7537e1ac560419))
+* let users report other users ([b12544f](https://adl.github.com/adl-developer/kinkane-backend/commit/b12544f93a793142a3c63cd0bd7537e1ac560419)) — [details](changelog/2026-07-13-report-user.md)
 
 
 ## 2026-07-10